requirements.txt
flask
pillow
transformers
torch
torchvision
torchaudio
groq
python-dotenv
requests